fix user delete

+++ controllers/userControl.js
@@ -22,21 +22,27 @@ module.exports = {
 		});
 	},
 	delete: (req, res) => {
-		Event.find({ userId: req.params.userId }).then((events) => {
-			events.forEach((event) => {
-				Expense.deleteMany({ eventId: event._id }).then((expenses) => {
-					console.log(expenses);
-					Event.deleteMany({ userId: req.params.userId }).then((events) => {
-						console.log(events);
-						Bill.deleteMany({ userId: req.params.userId }).then((bills) => {
-							console.log(bills);
-							User.deleteOne({ _id: req.params.userId }).then((user) => {
-								console.log(user);
+		User.findOne({ _id: req.params.userId }).then((user) => {
+			if (user.events) {
+				Event.find({ userId: req.params.userId }).then((events) => {
+					events.forEach((event) => {
+						Expense.deleteMany({ eventId: event._id }).then((expenses) => {
+							console.log(expenses);
+							Event.deleteMany({ userId: req.params.userId }).then((events) => {
+								console.log(events);
+								if (user.bills) {
+									Bill.deleteMany({ userId: req.params.userId }).then((bills) => {
+										console.log(bills);
+									});
+								}
+								User.deleteOne({ _id: req.params.userId }).then((user) => {
+									console.log(user);
+								});
 							});
 						});
 					});
 				});
-			});
+			}
 		});
 	}
 };
